susupay-backend/requirements.txt
# Web framework
fastapi==0.115.6
uvicorn[standard]==0.34.0

# Database
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.1

# Settings
pydantic-settings==2.7.1

# Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.2.1

# Async tasks
celery[redis]==5.4.0
redis==5.2.1

# AWS S3
boto3==1.36.2

# HTTP client
httpx==0.28.1

# File uploads
python-multipart==0.0.20

# Push notifications
pywebpush==2.0.1

# Image processing
Pillow==11.1.0

# PDF generation
weasyprint==63.1

# Testing
pytest==8.3.4
pytest-asyncio==0.25.0
pytest-cov==6.0.0
